refactor(tui): move Responses tests to owned seam
Move the existing 25-test Responses suite behind an external tests.rs seam while keeping the same client::responses::tests module path. Runtime code is unchanged.

This lowers owned production Rust from 645404 to 644417 lines while preserving the DeepSeek, Codex, streaming, retry, usage, and tool-schema coverage.
